requirements.txt
torch==2.0.1
torchvision==0.15.2
Pillow==9.5.0
gradio==3.14.0
httpx==0.23.0
